Refactor of Move refactoring
1) Expect/actual move checkbox is visible when expect/actual listed in selection list only
2) Fixed invalid temprary renaming (when source file name exists in target directory)
3) Fixed invalid file delition when "delete all files" checkbox is unchecked (KT-36072)
4) Removed UpdatePackageDirective checkbox (now it always updated instead of unpredicted behavior)
5) Removed SpecifyFileNameInPackage checkbox (in case of it redundancy)
6) Wrap resolve operations into the process
7) Minor refactorings
8) Fixed MPP declaration move top level declarations

+++ idea/src/org/jetbrains/kotlin/idea/refactoring/move/moveDeclarations/MoveKotlinDeclarationsProcessor.kt
@@ -363,7 +363,7 @@ class MoveKotlinDeclarationsProcessor(
                     }
                 }
 
-                if (descriptor.deleteSourceFiles) {
+                if (descriptor.deleteSourceFiles && sourceFile.declarations.isEmpty()) {
                     sourceFile.delete()
                 }
             }
